some cosmetic changes:
sides of the app are now different color than the rest of it

dropdown is now pointed downwards insted of leftwards. Color of it now darker and it has a bit of shadow.

example view now looks much better (shadow, cross is now reactive to hover)

previous commit accidentaly only pushed 1 file